Updating EEPROM code.

+++ Firmware/UnMannedParachuteController/UnMannedParachuteController/HAL/UART/ParseUart.hpp
@@ -401,7 +401,7 @@ class ParseGPSUart {
 			return result;		
 		}
 		
-		static double GetLatitude() {
+		static float GetLatitude() {
 			return latitude;
 		}
